CIR-1803: Revert emailv2service code as need to move the test responses to the verification service
