+++ test/uk/gov/hmrc/emailverification/services/EmailVerificationV2ServiceSpec.scala
@@ -67,7 +67,7 @@ class EmailVerificationV2ServiceSpec extends RepositoryBaseSpec {
 
     val mockAuditService: AuditV2Service = mock[AuditV2Service]
 
-    val mockEmailService: EmailV2Service = mock[EmailV2Service]
+    val mockEmailService: EmailService = mock[EmailService]
     val emailVerificationService = new EmailVerificationV2Service(mockVerificationCodeGenerator, verificationRepository, mockEmailService, mockAuditService)
   }
